requirements.txt
astroid==2.0.4
autopep8==1.4.4
awsebcli==3.16.0
beautifulsoup4==4.8.1
boto3==1.10.45
botocore==1.13.39
bs4==0.0.1
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.6
django-extensions==2.2.5
django-storages==1.8
djangorestframework==3.10.3
docutils==0.15.2
drf-yasg==1.17.0
future==0.16.0
geographiclib==1.50
geopy==1.20.0
httpie==1.0.3
idna==2.7
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.4
openapi-codec==1.3.2
packaging==19.2
pandas==0.25.3
pathspec==0.5.9
Pillow==6.2.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycodestyle==2.5.0
Pygments==2.4.2
pylint==2.1.1
pyparsing==2.4.5
pypiwin32==223
python-dateutil==2.8.0
python-decouple==3.3
pytz==2019.3
pywin32==227
PyYAML==3.13
requests==2.20.1
rope==0.11.0
rsa==3.4.2
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
semantic-version==2.5.0
simplejson==3.16.0
six==1.11.0
soupsieve==1.9.5
sqlparse==0.3.0
termcolor==1.1.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.7
wincertstore==0.2
wrapt==1.10.11
